[PATCH 1/1] rculist: Replaced list_first_entry_rcu() with list_first_or_null_rcu()
Replaced list_first_entry_rcu() with list_first_or_null_rcu() because

list_first_entry_rcu() is not safe as one can find in the comment that

this patch also adds.

This patch incorporated Paul's suggestions to the previous version of

this patch available here:

https://lkml.org/lkml/2012/4/2/536

This patch cannot break any upstream code because list_first_entry_rcu

is not being used anywhere in the kernel (tested with grep(1)), and

external code that uses it is probably broken already.

Signed-off-by: Michel Machado <michel [at] digirati>

CC: Paul E. McKenney <paulmck [at] linux>

CC: Dipankar Sarma <dipankar [at] in>

Signed-off-by: UupInTheAir <upintheair.xda@gmail.com>